requirements.txt
fastmcp
python-dotenv
pydantic
